roomImg 훅만들어서 사용함
