+++ src/hooks/useRoomImg.js
@@ -1,0 +1,109 @@
+import logo from '../component/bookpage/img/logo.png';
+import geomun from '../component/bookpage/img/geomun.jpeg';
+import darrenche from '../component/bookpage/img/darrenche.jpeg';
+import yongnuni from '../component/bookpage/img/yongnuni.jpeg';
+import darabi from '../component/bookpage/img/darabi.jpeg';
+import nabox from '../component/bookpage/img/nabox.jpeg';
+
+export default function useRoomImg() {
+  const roomsImg = [
+    {
+      room_id: 201,
+      img: logo,
+    },
+    {
+      room_id: 202,
+      img: logo,
+    },
+    {
+      room_id: 203,
+      img: logo,
+    },
+    {
+      room_id: 204,
+      img: logo,
+    },
+    {
+      room_id: 205,
+      img: logo,
+    },
+    {
+      room_id: 206,
+      img: logo,
+    },
+    {
+      room_id: 207,
+      img: logo,
+    },
+    {
+      room_id: 208,
+      img: logo,
+    },
+    {
+      room_id: 209,
+      img: logo,
+    },
+    {
+      room_id: 210,
+      img: logo,
+    },
+    {
+      room_id: 211,
+      img: logo,
+    },
+    {
+      room_id: 212,
+      img: nabox,
+    },
+    {
+      room_id: 213,
+      img: nabox,
+    },
+    {
+      room_id: 214,
+      img: nabox,
+    },
+    {
+      room_id: 215,
+      img: nabox,
+    },
+    {
+      room_id: 216,
+      img: nabox,
+    },
+    {
+      room_id: 217,
+      img: nabox,
+    },
+    {
+      room_id: 301,
+      img: darrenche,
+    },
+    {
+      room_id: 302,
+      img: yongnuni,
+    },
+    {
+      room_id: 303,
+      img: darabi,
+    },
+    {
+      room_id: 304,
+      img: geomun,
+    },
+    {
+      room_id: 305,
+      img: nabox,
+    },
+    {
+      room_id: 306,
+      img: nabox,
+    },
+    {
+      room_id: 307,
+      img: nabox,
+    },
+  ];
+
+  return roomsImg;
+}
